+++ Client/Frontend/Browser/DownloadQueue.swift
@@ -68,12 +68,6 @@ class HTTPDownload: Download {
 
     private var resumeData: Data?
 
-    // Used to avoid name spoofing using Unicode RTL char to change file extension
-    public static func stripUnicode(fromFilename string: String) -> String {
-        let allowed = CharacterSet.alphanumerics.union(CharacterSet.punctuationCharacters)
-        return string.components(separatedBy: allowed.inverted).joined()
-     }
-
     init(preflightResponse: URLResponse, request: URLRequest) {
         self.preflightResponse = preflightResponse
         self.request = request
@@ -81,7 +75,7 @@ class HTTPDownload: Download {
         super.init()
 
         if let filename = preflightResponse.suggestedFilename {
-            self.filename = HTTPDownload.stripUnicode(fromFilename: filename)
+            self.filename = filename
         }
 
         if let mimeType = preflightResponse.mimeType {
