+++ ClientTests/StringExtensionsTests.swift
@@ -4,7 +4,6 @@
 
 import Foundation
 import XCTest
-@testable import Client
 
 class StringExtensionsTests: XCTestCase {
 
@@ -52,11 +51,4 @@ class StringExtensionsTests: XCTestCase {
         roundtripTest("http://mozilla.com/?a=foo&b=bar", "http://mozilla.com/%3Fa%3Dfoo%26b%3Dbar")
     }
 
-    func testRemoveUnicodeFromFilename() {
-        let file = "foo-\u{200F}cod.jpg" // Unicode RTL-switch code, becomes "foo-gpj.doc"
-        let nounicode = "foo-cod.jpg"
-        XCTAssert(file != nounicode)
-        let strip = HTTPDownload.stripUnicode(fromFilename: file)
-        XCTAssert(strip == nounicode)
-    }
 }
