--- Extensions/Today/UIButtonExtensions.swift
@@ -1,21 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-import UIKit
-
-extension UIButton {
-    func setBackgroundColor(_ color: UIColor, forState state: UIControl.State) {
-        let colorView = UIView(frame: CGRect(width: 1, height: 1))
-        colorView.backgroundColor = color
-
-        UIGraphicsBeginImageContext(colorView.bounds.size)
-        if let context = UIGraphicsGetCurrentContext() {
-            colorView.layer.render(in: context)
-        }
-        let colorImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-
-        self.setBackgroundImage(colorImage, for: state)
-    }
-}
